Add various bash scripts for advanced training and model evaluation
Added bash scripts for various operations including advanced training of models on Hecktor dataset, model checkpoint verification, model evaluation on adversarial images, and parameter visualization. Elements of the graphics and model processing functionality are revised and improved. The new scripts are designed to enhance the efficiency of model training and evaluation workflows.